requirements.txt
# Core packages
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
streamlit>=1.8.0

# Machine Learning & Statistics
scikit-learn>=1.0.0
joblib>=1.1.0
shap>=0.40.0
statsmodels>=0.13.2

# Data processing
requests>=2.26.0
python-dotenv>=0.19.0
rasterio>=1.2.0
geopandas>=0.10.0

# Testing
pytest>=6.2.0

# Documentation
sphinx>=4.2.0